+++ lib/src/rendering/sliver_gap.dart
@@ -1,0 +1,74 @@
+import 'package:flutter/rendering.dart';
+
+class RenderSliverGap extends RenderSliver {
+  RenderSliverGap({
+    double mainAxisExtent,
+    Color color,
+  })  : _mainAxisExtent = mainAxisExtent,
+        _color = color;
+
+  double get mainAxisExtent => _mainAxisExtent;
+  double _mainAxisExtent;
+  set mainAxisExtent(double value) {
+    if (_mainAxisExtent != value) {
+      _mainAxisExtent = value;
+      markNeedsLayout();
+    }
+  }
+
+  Color get color => _color;
+  Color _color;
+  set color(Color value) {
+    if (_color != value) {
+      _color = value;
+      markNeedsPaint();
+    }
+  }
+
+  @override
+  void performLayout() {
+    final double paintExtent = calculatePaintOffset(
+      constraints,
+      from: 0,
+      to: mainAxisExtent,
+    );
+    final double cacheExtent = calculateCacheOffset(
+      constraints,
+      from: 0,
+      to: mainAxisExtent,
+    );
+
+    assert(paintExtent.isFinite);
+    assert(paintExtent >= 0.0);
+    geometry = SliverGeometry(
+      scrollExtent: mainAxisExtent,
+      paintExtent: paintExtent,
+      cacheExtent: cacheExtent,
+      maxPaintExtent: mainAxisExtent,
+      hitTestExtent: paintExtent,
+      hasVisualOverflow: mainAxisExtent > constraints.remainingPaintExtent ||
+          constraints.scrollOffset > 0.0,
+    );
+  }
+
+  @override
+  void paint(PaintingContext context, Offset offset) {
+    if (color != null) {
+      final Paint paint = Paint()..color = color;
+      final Size size = constraints
+          .asBoxConstraints(
+            minExtent: _mainAxisExtent,
+            maxExtent: _mainAxisExtent,
+          )
+          .constrain(Size.zero);
+      context.canvas.drawRect(offset & size, paint);
+    }
+  }
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(DoubleProperty('mainAxisExtent', mainAxisExtent));
+    properties.add(ColorProperty('color', color));
+  }
+}