+++ lib/src/widgets/gap.dart
@@ -1,0 +1,172 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
+import 'package:gap/src/rendering/gap.dart';
+
+/// A widget that takes a fixed amount of space in a [Row], [Column],
+/// or [Flex] widget.
+///
+/// A [Gap] widget must be a descendant of a [Row], [Column], or [Flex],
+/// and the path from the [Gap] widget to its enclosing [Row], [Column], or
+/// [Flex] must contain only [StatelessWidget]s or [StatefulWidget]s (not other
+/// kinds of widgets, like [RenderObjectWidget]s).
+///
+/// See also:
+///
+///  * [MaxGap], a gap that can take, at most, the amount of space specified.
+///  * [SliverGap], the sliver version of this widget.
+class Gap extends LeafRenderObjectWidget {
+  /// Creates a widget that takes a fixed [mainAxisExtent] of space in
+  /// a [Row], [Column], or [Flex] widget.
+  ///
+  /// The [mainAxisExtent] must not be null and must be positive.
+  /// The [crossAxisExtent] must be either null or positive.
+  const Gap(
+    this.mainAxisExtent, {
+    Key key,
+    this.crossAxisExtent,
+    this.color,
+  })  : assert(mainAxisExtent != null &&
+            mainAxisExtent >= 0 &&
+            mainAxisExtent < double.infinity),
+        assert(crossAxisExtent == null || crossAxisExtent >= 0),
+        super(key: key);
+
+  /// Creates a widget that takes a fixed [mainAxisExtent] of space in
+  /// a [Row], [Column], or [Flex] widget and expands in the cross axis
+  /// direction.
+  ///
+  /// The [mainAxisExtent] must not be null and must be positive.
+  const Gap.expand(
+    double mainAxisExtent, {
+    Key key,
+    Color color,
+  }) : this(
+          mainAxisExtent,
+          key: key,
+          crossAxisExtent: double.infinity,
+          color: color,
+        );
+
+  /// The amount of space this widget takes in the direction of the parent.
+  ///
+  /// If the parent is a [Column] this is the height of this widget.
+  /// If the parent is a [Row] this is the width of this widget.
+  ///
+  /// Must not be null and must be positive.
+  final double mainAxisExtent;
+
+  /// The amount of space this widget takes in the opposite direction of the
+  /// parent.
+  ///
+  /// If the parent is a [Column] this is the width of this widget.
+  /// If the parent is a [Row] this is the height of this widget.
+  ///
+  /// Must be positive or null. If it's null (the default) the cross axis extent
+  /// will be the same as the constraints of the parent in the opposite
+  /// direction.
+  final double crossAxisExtent;
+
+  /// The color used to fill the gap.
+  final Color color;
+
+  @override
+  RenderObject createRenderObject(BuildContext context) {
+    return RenderGap(
+      mainAxisExtent: mainAxisExtent,
+      crossAxisExtent: crossAxisExtent ?? 0,
+      color: color,
+    );
+  }
+
+  @override
+  void updateRenderObject(BuildContext context, RenderGap renderObject) {
+    renderObject
+      ..mainAxisExtent = mainAxisExtent
+      ..crossAxisExtent = crossAxisExtent ?? 0
+      ..color = color;
+  }
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(DoubleProperty('mainAxisExtent', mainAxisExtent));
+    properties.add(
+        DoubleProperty('crossAxisExtent', crossAxisExtent, defaultValue: 0));
+    properties.add(ColorProperty('color', color));
+  }
+}
+
+/// A widget that takes, at most, an amount of space in a [Row], [Column],
+/// or [Flex] widget.
+///
+/// A [MaxGap] widget must be a descendant of a [Row], [Column], or [Flex],
+/// and the path from the [MaxGap] widget to its enclosing [Row], [Column], or
+/// [Flex] must contain only [StatelessWidget]s or [StatefulWidget]s (not other
+/// kinds of widgets, like [RenderObjectWidget]s).
+///
+/// See also:
+///
+///  * [Gap], the unflexible version of this widget.
+class MaxGap extends StatelessWidget {
+  /// Creates a widget that takes, at most, the specified [mainAxisExtent] of
+  /// space in a [Row], [Column], or [Flex] widget.
+  ///
+  /// The [mainAxisExtent] must not be null and must be positive.
+  /// The [crossAxisExtent] must be either null or positive.
+  const MaxGap(
+    this.mainAxisExtent, {
+    Key key,
+    this.crossAxisExtent,
+    this.color,
+  }) : super(key: key);
+
+  /// Creates a widget that takes, at most, the specified [mainAxisExtent] of
+  /// space in a [Row], [Column], or [Flex] widget and expands in the cross axis
+  /// direction.
+  ///
+  /// The [mainAxisExtent] must not be null and must be positive.
+  /// The [crossAxisExtent] must be either null or positive.
+  const MaxGap.expand(
+    double mainAxisExtent, {
+    Key key,
+    Color color,
+  }) : this(
+          mainAxisExtent,
+          key: key,
+          crossAxisExtent: double.infinity,
+          color: color,
+        );
+
+  /// The amount of space this widget takes in the direction of the parent.
+  ///
+  /// If the parent is a [Column] this is the height of this widget.
+  /// If the parent is a [Row] this is the width of this widget.
+  ///
+  /// Must not be null and must be positive.
+  final double mainAxisExtent;
+
+  /// The amount of space this widget takes in the opposite direction of the
+  /// parent.
+  ///
+  /// If the parent is a [Column] this is the width of this widget.
+  /// If the parent is a [Row] this is the height of this widget.
+  ///
+  /// Must be positive or null. If it's null (the default) the cross axis extent
+  /// will be the same as the constraints of the parent in the opposite
+  /// direction.
+  final double crossAxisExtent;
+
+  /// The color used to fill the gap.
+  final Color color;
+
+  @override
+  Widget build(BuildContext context) {
+    return Flexible(
+      child: Gap(
+        mainAxisExtent,
+        crossAxisExtent: crossAxisExtent,
+        color: color,
+      ),
+    );
+  }
+}