+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+## 1.1.0
+### Added
+- `SliverGap` the sliver version of `Gap`.
+
 ## 1.0.1
-* Fix debug properties in Gap widget
+### Fixed
+- Fix debug properties in Gap widget
 
 ## 1.0.0
-* Initial Open Source release
+- Initial Open Source release